+++ LogTap-Noop/src/main/java/com/github/husseinhj/logtap/LogTapInterceptor.kt
@@ -1,8 +1,0 @@
-package com.github.husseinhj.logtap
-import okhttp3.Interceptor
-import okhttp3.Response
-
-class LogTapInterceptor : Interceptor {
-  override fun intercept(chain: Interceptor.Chain): Response =
-    chain.proceed(chain.request())
-}