+++ LogTap-Noop/src/main/java/com/github/husseinhj/logtap/LogTapSinkAdapter.kt
@@ -1,5 +1,0 @@
-package com.github.husseinhj.logtap
-
-class LogTapSinkAdapter: LogTapLogcatBridge.Sink {
-    override fun onLog(priority: Char, tag: String, message: String, threadId: Int?, time: String?) {}
-}